take into consideration no db case

+++ frontend/ui/data/onetime_migration.lua
@@ -6,6 +6,7 @@ local DataStorage = require("datastorage")
 local lfs = require("libs/libkoreader-lfs")
 local logger = require("logger")
 local SQ3 = require("lua-ljsqlite3/init")
+local util = require("util")
 local _ = require("gettext")
 
 -- Date at which the last migration snippet was added
@@ -589,17 +590,22 @@ end
 -- 20230727, Statistics plugin null id_book in page_stat_data
 if last_migration_date < 20230727 then
     logger.info("Performing one-time migration for 20230727")
-    local conn = SQ3.open(DataStorage:getSettingsDir() .. "/statistics.sqlite3")
-    local ok, value = pcall(conn.exec, conn, "PRAGMA table_info('page_stat_data')")
-    if ok and value then
-        -- Has table
-        conn:exec("DELETE FROM page_stat_data WHERE id_book IS null;")
-        local ok2, errmsg = pcall(conn.exec, conn, "VACUUM;")
-        if not ok2 then
-            logger.warn("Failed compacting statistics database when fixing null id_book:", errmsg)
+    local db_location = DataStorage:getSettingsDir() .. "/statistics.sqlite3"
+    if util.fileExists(db_location) then
+        local conn = SQ3.open(db_location)
+        local ok, value = pcall(conn.exec, conn, "PRAGMA table_info('page_stat_data')")
+        if ok and value then
+            -- Has table
+            conn:exec("DELETE FROM page_stat_data WHERE id_book IS null;")
+            local ok2, errmsg = pcall(conn.exec, conn, "VACUUM;")
+            if not ok2 then
+                logger.warn("Failed compacting statistics database when fixing null id_book:", errmsg)
+            end
+        else
+            logger.warn("db not compatible when performing onetime migration:", ok, value)
         end
     else
-        logger.warn("db not compatible when performing onetime migration:", ok, value)
+        logger.info("statistics.sqlite3 not found.")
     end
 end
